Move navbar directory to component
