+++ my-app/pages/index.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from "next";
 import Head from "next/head";
-import Navbar from "./navbar";
+import Navbar from "../components/navbar";
 
 const Home: NextPage = () => {
   return (
